Linking in more content to ToC (#24)
Two useful articles from @HumairAK 

I'm not sure this is the best way for me to link these in, but it works. I think we need to redo how the ToC works when it goes to the website and find a way to make it clear how to crosslink files from different repos. Maybe there is a GH org trick we can use?